Kept only XCTestCase that reuses same session for all tests
